requirements.txt

torch
torchvision
pandas
Pillow
tensorboard
